01_Read_from_TTN/mqtt/requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.0.23
paho-mqtt==1.5.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
wrapt==1.11.2
